requirements.txt
streamlit==1.37.0
Pillow==10.4.0
numpy==1.26.4
reportlab==4.2.2
fonttools==4.46.3